requirements.txt
click==6.6
Django==1.10
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
numpy==1.11.1
peewee==2.8.2
Pillow==3.3.0
python-crontab==2.1.1
python-dateutil==2.5.3
sense-hat==2.2.0
six==1.10.0
Werkzeug==0.11.10
